removing pull-up/pull-down works! experiment by setting to active low
the pull-resistors are too small compared to the 10kOhm in the encoder circuit, pull-resistor and 10kOhm on the pcb form a voltage divider. That in the case of super mini nrf "sees" a voltage of 1.38V when encoder is turned instead of 0V/GND. We have the same issue with QMK on the RP2040, but luckily there the pull resistor is higher, thus the PIN "sees" 0.56V on the RP2040 which is low enough to be registered as "low".